# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Monitoring Platform API",
    "description": "API for system monitoring, alerts, notifications, and reporting",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/api/integration/widgets/validate": {
      "post": {
        "tags": [
          "Integration"
        ],
        "summary": "Validate widget source",
        "description": "Validate third-party widget script URL",
        "operationId": "validateWidgetSource",
        "parameters": [
          {
            "name": "scriptCode",
            "in": "query",
            "description": "Script tag HTML",
            "required": false,
            "schema": {
              "type": "string",
              "default": "<script src=\"https://cdn.example.com/widget.js\"></script>"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/widgets/secureValidate": {
      "post": {
        "tags": [
          "Integration"
        ],
        "summary": "Validate widget source (secure)",
        "description": "Validate widget with domain whitelist",
        "operationId": "secureValidateWidgetSource",
        "parameters": [
          {
            "name": "scriptCode",
            "in": "query",
            "description": "Script tag HTML",
            "required": false,
            "schema": {
              "type": "string",
              "default": "<script src=\"https://cdn.example.com/widget.js\"></script>"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/integration/services/test": {
      "get": {
        "tags": [
          "Integration"
        ],
        "summary": "Test service connection",
        "description": "Test connectivity to external service",
        "operationId": "testServiceConnection",
        "parameters": [
          {
            "name": "serviceUrl",
            "in": "query",
            "description": "Service URL",
            "required": false,
            "schema": {
              "type": "string",
              "default": "https://api.example.com/health"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/integration/services/secureTest": {
      "get": {
        "tags": [
          "Integration"
        ],
        "summary": "Test service connection (secure)",
        "description": "Test connectivity with validation",
        "operationId": "secureTestServiceConnection",
        "parameters": [
          {
            "name": "serviceUrl",
            "in": "query",
            "description": "Service URL",
            "required": false,
            "schema": {
              "type": "string",
              "default": "https://api.example.com/health"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}